ai-platform-unified/tutorials/unofficial/tf-keras-image-classification-custom-train-multi-replica/trainer/requirements.txt
tensorflow==2.4.1
google-cloud-storage==1.35.0
tensorflow_datasets
